Infrared in debug options
