+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/DebugCategory.kt
@@ -96,4 +96,10 @@ private fun DebugCategoryItems(
         state = settings.selfUpdaterDebug,
         onSwitchState = debugViewModel::onSwitchSelfUpdaterDebug
     )
+    GrayDivider()
+    SwitchableElement(
+        titleId = R.string.experimental_new_infrared,
+        state = settings.useNewInfrared,
+        onSwitchState = debugViewModel::onSwitchNewInfrared
+    )
 }
