+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/ExperimentalCategory.kt
@@ -8,7 +8,6 @@ import com.flipperdevices.settings.impl.R
 import com.flipperdevices.settings.impl.composable.elements.CategoryElement
 import com.flipperdevices.settings.impl.composable.elements.ClickableElement
 import com.flipperdevices.settings.impl.composable.elements.GrayDivider
-import com.flipperdevices.settings.impl.composable.elements.SwitchableElement
 import com.flipperdevices.settings.impl.viewmodels.ExperimentalViewModel
 import com.flipperdevices.settings.impl.viewmodels.SettingsViewModel
 import tangle.viewmodel.compose.tangleViewModel
@@ -41,12 +40,6 @@ fun ExperimentalCategory(
                 onClick = {
                     experimentalViewModel.onOpenScreenStreaming(navController)
                 }
-            )
-            GrayDivider()
-            SwitchableElement(
-                titleId = R.string.experimental_new_infrared,
-                state = settings.useNewInfrared,
-                onSwitchState = experimentalViewModel::onSwitchNewInfrared
             )
         }
     }
