+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -1,23 +1,18 @@
 package com.flipperdevices.settings.impl.viewmodels
 
-import androidx.datastore.core.DataStore
-import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.flipperdevices.core.di.provideDelegate
-import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
 import com.flipperdevices.settings.api.SettingsFeatureEntry
-import kotlinx.coroutines.launch
 import tangle.viewmodel.VMInject
 import javax.inject.Provider
 
 class ExperimentalViewModel @VMInject constructor(
     fileManagerEntryProvider: Provider<FileManagerEntry>,
     metricApiProvider: Provider<MetricApi>,
-    private val settingsDataStore: DataStore<Settings>,
     private val settingsFeatureEntry: SettingsFeatureEntry
 ) : LifecycleViewModel() {
     private val fileManagerEntry by fileManagerEntryProvider
@@ -31,15 +26,5 @@ class ExperimentalViewModel @VMInject constructor(
     fun onOpenScreenStreaming(navController: NavController) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_SCREEN_STREAMING)
         navController.navigate(settingsFeatureEntry.remoteControlStub())
-    }
-
-    fun onSwitchNewInfrared(enabled: Boolean) {
-        viewModelScope.launch {
-            settingsDataStore.updateData {
-                it.toBuilder()
-                    .setUseNewInfrared(enabled)
-                    .build()
-            }
-        }
     }
 }
